requirements.txt
django
djangorestframework
python-decouple
djangorestframework-simplejwt
pillow
phonenumbers
django-jazzmin
drf-material
icecream
uvicorn
gunicorn
whitenoise
django-debug-toolbar
django-cors-headers
django-extensions
django-import-export[all]
environs
requests